save current work state

+++ controller/calendarcontroller.php
@@ -7,20 +7,19 @@
  */
 namespace OCA\Calendar\Controller;
 
-use \OCA\Calendar\AppFramework\Core\API;
-use \OCA\Calendar\AppFramework\Http\Http;
-use \OCA\Calendar\AppFramework\Http\Request;
+use \OCP\AppFramework\Http\Http;
 
-use \OCA\Calendar\AppFramework\DoesNotExistException;
-
-use \OCA\Calendar\BusinessLayer\CalendarBusinessLayer;
-use \OCA\Calendar\BusinessLayer\ObjectBusinessLayer;
+use \OCA\Calendar\Db\DoesNotExistException;
 use \OCA\Calendar\BusinessLayer\BusinessLayerException;
 
 use \OCA\Calendar\Db\Calendar;
 use \OCA\Calendar\Db\CalendarCollection;
 
 use \OCA\Calendar\Http\JSONResponse;
+
+use \OCA\Calendar\Http\ICS\ICSCalendar;
+use \OCA\Calendar\Http\ICS\ICSCalendarCollection;
+use \OCA\Calendar\Http\ICS\ICSCalendarReader;
 
 use \OCA\Calendar\Http\JSON\JSONCalendar;
 use \OCA\Calendar\Http\JSON\JSONCalendarCollection;
@@ -40,12 +39,18 @@ class CalendarController extends Controller {
 			$limit = $this->header('X-OC-CAL-LIMIT', 'integer');
 			$offset	= $this->header('X-OC-CAL-OFFSET', 'integer');
 
-			$calendarCollection = $this->calendarBusinessLayer->findAll($userId, $limit, $offset);
-			$jsonCalendarCollection = new JSONCalendarCollection($calendarCollection);
+			$doesAcceptRawICS = $this->doesClientAcceptRawICS();
 
-			return new JSONResponse($jsonCalendarCollection->serialize());
+			$calendarCollection = $this->calendarBusinessLayer->findAll($userId, $limit, $offset);
+			if($doesAcceptRawICS === true) {
+				$serializer = new ICSCalendarCollection($calendarCollection);
+			} else {
+				$serializer = new JSONCalendarCollection($calendarCollection);
+			}
+
+			return new JSONResponse($serializer);
 		} catch (BusinessLayerException $ex) {
-			$this->api->log($ex->getMessage(), 'warn');
+			$this->api->log($ex->getMessage(), 'debug');
 			return new JSONResponse(null, HTTP::STATUS_BAD_REQUEST);
 		}
 	}
@@ -61,17 +66,18 @@ class CalendarController extends Controller {
 			$userId = $this->api->getUserId();
 			$calendarId = $this->params('calendarId');
 
+			$doesAcceptRawICS = $this->doesClientAcceptRawICS();
+
 			$calendar = $this->calendarBusinessLayer->find($calendarId, $userId);
+			if($doesAcceptRawICS === true) {
+				$serializer = new ICSCalendar($calendarCollection);
+			} else {
+				$serializer = new JSONCalendar($calendarCollection);
+			}
 
-			$jsonCalendar = new JSONCalendar($calendar);
-
-			$response = new Reponse();
-			$response->addHeader('X-Content-Type-Options', 'nosniff');
-			$response->addHeader('Content-type', 'application/json; charset=utf-8');
-
-			return new JSONResponse($jsonCalendar->serialize());
+			return new JSONResponse($serializer);
 		} catch (BusinessLayerException $ex) {
-			$this->api->log($ex->getMessage(), 'warn');
+			$this->api->log($ex->getMessage(), 'debug');
 			return new JSONResponse(null, HTTP::STATUS_BAD_REQUEST);
 		}
 	}
@@ -85,19 +91,31 @@ class CalendarController extends Controller {
 	public function create() {
 		try {
 			$userId = $this->api->getUserId();
-			$json = $this->request->params;
+			$data = $this->request->params;
 
-			$jsonReader = new JSONCalendarReader($json);
-			$calendar = $jsonReader->getCalendar()
-								   ->setUserId($userId)
-								   ->setOwnerId($userId);
+			$didSendRawICS = $this->didClientSendRawICS();
+			$doesAcceptRawICS = $this->doesClientAcceptRawICS();
 
-			$calendar = $this->calendarBusinessLayer->create($calendar, $userId);
-			$jsonCalendar = new JSONCalendar($calendar);
+			if($didSendRawICS === true) {
+				$reader = new ICSCalendarReader($data);
+			} else {
+				$reader = new JSONCalendarReader($data);
+			}
 
-			return new JSONResponse($jsonCalendar->serialize(), HTTP::STATUS_CREATED);
+			$calendar = $reader->sanitize()->getCalendar();
+			$calendar->setUser($userId)->setOwner($userId);
+
+			$calendar = $this->calendarBusinessLayer->create($calendar);
+
+			if($doesAcceptRawICS === true) {
+				$serializer = new ICSCalendar($calendar);
+			} else {
+				$serializer = new JSONCalendar($calendar);
+			}
+
+			return new JSONResponse($serializer, HTTP::STATUS_CREATED);
 		} catch (BusinessLayerException $ex) {
-			$this->api->log($ex->getMessage(), 'warn');
+			$this->api->log($ex->getMessage(), 'debug');
 			return new JSONResponse(null, HTTP::STATUS_BAD_REQUEST);
 		}
 	}
@@ -112,18 +130,36 @@ class CalendarController extends Controller {
 		try {
 			$userId = $this->api->getUserId();
 			$calendarId = $this->params('calendarId');
-			$json = $this->request->params;
+			$data = $this->request->params;
 
-			$jsonReader = new JSONCalendarReader($json);
-			$calendar = $jsonReader->getCalendar()
-								   ->setUserId($userId);
+			$didSendRawICS = $this->didClientSendRawICS();
+			$doesAcceptRawICS = $this->doesClientAcceptRawICS();
+
+			//check if calendar exists, if not return 404
+			if($this->calendarBusinessLayer->doesExist($calendarId, $userId) === false) {
+				return new JSONResponse(null, HTTP::STATUS_NOT_FOUND);
+			}
+
+			if($didSendRawICS === true) {
+				$reader = new ICSCalendarReader($data);
+			} else {
+				$reader = new JSONCalendarReader($data);
+			}
+
+			$calendar = $reader->sanitize()->getCalendar();
+			$calendar->setUser($userId);
 
 			$calendar = $this->calendarBusinessLayer->update($calendar, $calendarId, $userId);
-			$jsonCalendar = new JSONCalendar($calendar);
 
-			return new JSONResponse($jsonCalendar->serialize(), HTTP::STATUS_CREATED);
+			if($doesAcceptRawICS === true) {
+				$serializer = new ICSCalendar($calendar);
+			} else {
+				$serializer = new JSONCalendar($calendar);
+			}
+
+			return new JSONResponse($serializer);
 		} catch(BusinessLayerException $ex) {
-			$this->api->log($ex->getMessage(), 'warn');
+			$this->api->log($ex->getMessage(), 'debug');
 			return new JSONResponse(null, HTTP::STATUS_BAD_REQUEST);
 		}
 	}
@@ -139,9 +175,14 @@ class CalendarController extends Controller {
 			$userId	= $this->api->getUserId();
 			$calendarId	= $this->params('calendarId');
 
+			//check if calendar exists, if not return 404
+			if($this->calendarBusinessLayer->doesExist($calendarId, $userId) === false) {
+				return new JSONResponse(null, HTTP::STATUS_NOT_FOUND);
+			}
+
 			$this->calendarBusinessLayer->delete($calendarId, $userId);
 
-			return new JSONResponse(null, HTTP::STATUS_NO_CONTENT);
+			return new JSONResponse();
 		} catch (BusinessLayerException $ex) {
 			$this->api->log($ex->getMessage(), 'warn');
 			return new JSONResponse(null, HTTP::STATUS_BAD_REQUEST);
